Created folder in components folder and added dashboard components in there
