+++ client/app/utils/colors.ts
@@ -1,0 +1,8 @@
+const colorPalette = {
+    light: "#eeeeee",
+    medium: "#323232",
+    dark: "#181818",
+    green: "#28ce78",
+};
+
+export default colorPalette;
