+++ client/app/utils/dashboardContainerBoxShadow.ts
@@ -1,0 +1,1 @@
+export const boxShadowForContainers = " rgba(0, 0, 0, 0.16) 0px 3px 6px, rgba(0, 0, 0, 0.23) 0px 3px 6px"
